+++ datconv_pkg/readers/dccsv.py
@@ -135,7 +135,11 @@ class DCReader:
                     self._wri.writeRecord(rec)
 
         # OBLIGATORY
-        self._wri.writeFooter([])
+        footer = []
+        if self._flt is not None:
+            if hasattr(self._flt, 'setFooter'):
+                self._flt.setFooter(footer)
+        self._wri.writeFooter(footer)
         
         fout.close()
         Log.info('Output saved to %s' % outpath)